CommandUtilsの追加と関連する関数の実装
